+++ src/main/java/com/facedynamics/notifications/services/NotificationService.java
@@ -1,19 +1,19 @@
 package com.facedynamics.notifications.services;
 
 import com.facedynamics.notifications.model.Notification;
-import com.facedynamics.notifications.model.dto.NotificationGetDTO;
-import com.facedynamics.notifications.model.dto.NotificationResponseDTO;
+import com.facedynamics.notifications.model.NotificationResponseDTO;
+import com.facedynamics.notifications.model.dto.NotificationDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 
 public interface NotificationService {
 
-    Page<Notification> getAllNotificationsByUserId(int ownerId, Pageable pageable);
+    Page<Notification> getAllNotificationsByUserId(Long ownerId, Pageable pageable);
 
-    void deleteAllNotificationsByOwnerId(int ownerId);
+    void deleteAllNotificationsByOwnerId(Long ownerId);
 
-    Long deleteNotificationById(long notificationId);
+    Long deleteNotificationById(Long notificationId);
 
     NotificationResponseDTO createNotification(NotificationDto notification);
