+++ src/main/java/com/facedynamics/notifications/services/NotificationServiceImpl.java
@@ -14,10 +14,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
-import static com.facedynamics.notifications.utils.Constants.*;
+import static com.facedynamics.notifications.utils.Constants.ID_OF_THE_NOTIFICATION_IS_NOT_FOUND;
+import static com.facedynamics.notifications.utils.Constants.USER_IS_NOT_FOUND;
 
 /**
  * Service class that is responsible for business logic of
@@ -46,7 +46,7 @@ public class NotificationServiceImpl implements NotificationService {
      * if ID of the user is not found in the database.
      */
     @Override
-    public Page<Notification> getAllNotificationsByUserId(int ownerId, Pageable pageable) {
+    public Page<Notification> getAllNotificationsByUserId(Long ownerId, Pageable pageable) {
         if (notificationRepository.existsByOwnerId(ownerId)) {
             return notificationRepository.findAllByOwnerId(ownerId, pageable);
         } else {
@@ -60,7 +60,7 @@ public class NotificationServiceImpl implements NotificationService {
      * @param ownerId ID of the user who owns his notifications
      */
     @Override
-    public void deleteAllNotificationsByOwnerId(int ownerId) {
+    public void deleteAllNotificationsByOwnerId(Long ownerId) {
         if (notificationRepository.existsByOwnerId(ownerId)) {
             notificationRepository.deleteNotificationsByOwnerId(ownerId);
         } else {
@@ -77,7 +77,7 @@ public class NotificationServiceImpl implements NotificationService {
      * found in the database.
      */
     @Override
-    public Long deleteNotificationById(long Id) {
+    public Long deleteNotificationById(Long Id) {
         Optional<Notification> notification = notificationRepository.findNotificationById(Id);
         if (notification.isEmpty()) {
             throw new NotFoundException(ID_OF_THE_NOTIFICATION_IS_NOT_FOUND, Id);
